requirements.txt
amqp==2.6.1
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
cssselect==1.1.0
Django==3.1.4
django-cors-headers==3.5.0
django-rest-framework-nested==0.0.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
drf-nested-routers==0.91
drf-yasg==1.20.0
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
inflection==0.5.1
itemadapter==0.2.0
itemloaders==1.0.4
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.11
lxml==4.6.2
MarkupSafe==1.1.1
packaging==20.7
parsel==1.6.0
Pillow==8.0.1
Protego==0.1.16
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
pyparsing==2.4.7
python-scrapyd-api==2.1.2
pytz==2020.4
queuelib==1.5.0
redis==3.4.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Scrapy==2.4.1
scrapyd==1.2.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
uritemplate==3.0.1
urllib3==1.26.2
vine==1.3.0
w3lib==1.22.0
zope.interface==5.2.0
